Reland "fix up supported devices page"
