+++ .gitignore
@@ -8,8 +8,6 @@ node_modules
 .DS_Store
 
 # Will be initilized on run/build time
-src/docs/firmware/supported-devices.md
-src/.vuepress/public/supported-devices.js
 src/.vuepress/public/devices.json
 
 # old ignored stuff
